Made TupleBPS use a stringstore based Row for scratch space instead of a linear Row.  Also made the offsets arrays in RowGroup 64-bits. It builds here.  Untested.

+++ dbcon/joblist/groupconcat.cpp
@@ -205,7 +205,7 @@ void GroupConcatInfo::mapColumns(const RowGroup& projRG)
 
     for (vector<SP_GroupConcat>::iterator k = fGroupConcat.begin(); k != fGroupConcat.end(); k++)
     {
-        vector<uint32_t> pos;
+        vector<uint64_t> pos;
         vector<uint32_t> oids;
         vector<uint32_t> keys;
         vector<uint32_t> scale;
